backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-dateutil
plotly
pandas
boto3
sentence-transformers
pinecone-client
langgraph
langchain
tiktoken
litellm
google-generativeai
openai
anthropic
snowflake-connector-python
matplotlib
seaborn
wordcloud
serpapi
